Fix migration
The custom extension `install` for version 2 only installed the
multiplication function and operators, but didn't add anything else, not
even then `money_with_currency` type, which is required for the
multiplication operators and and everything else.

With a new version 3, we install the `money_with_currency` type, then
everything else, and uninstall each piece in the strict opposite order.